requirements.txt
# --- Web API 服务框架 ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# --- 大模型与编排 (LangChain 全家桶) ---
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.33
# 计算 Token 数量所需
tiktoken>=0.6.0

# --- 向量数据库 (本地存储) ---
chromadb>=0.4.24

# --- 科学计算与匹配算法 (用于余弦相似度) ---
scikit-learn>=1.4.2
numpy>=1.26.0
pandas>=2.2.0

# --- 爬虫与文档处理 ---
requests>=2.31.0
beautifulsoup4>=4.12.3
# 用于读取 PDF (RAG 必备)
pypdf>=4.2.0

# --- 数据库驱动 ---
pymysql>=1.1.0
sqlalchemy>=2.0.0

# --- 环境变量管理 ---
python-dotenv>=1.0.1

# --- 安全与加密 ---
cryptography>=41.0.4